# swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Mate Exam API"
  description: "A simple API to give test exams to students.
  Flow for an exam is like so:
  1. Pick a student.
  2. Pick an exam.
  3. Create result for chosen student and exam.
  4. Answer questions.
  5. Finish result."
host: "agentmate.herokuapp.com"
basePath: "/api/v1"
...
schemes:
- "https"
paths:
  /students:
    get:
      tags:
      - "students"
      summary: "Get list of all students"
      operationId: "getStudents"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Student"
    post:
      tags:
      - "students"
      summary: "Create a new student"
      operationId: "createStudent"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "body"
        in: "body"
        description: "Student object that needs to be created. _id does not need to be provided. If not provided it will be generated server side and returned in the response."
        required: true
        schema:
          $ref: "#/definitions/Student"
      responses:
            200:
              description: "successful operation"
              schema:
                $ref: "#/definitions/Student"
  /students/{id}:
    get:
      tags:
      - "students"
      summary: "Get a student by id"
      operationId: "getStudentById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of student to return"
        required: true
        type: "string"
        format: "hexadecimal"
        example: "5ad289124c9956964c868f2b"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Student"
    put:
      tags:
      - "students"
      summary: "Update a student by id"
      operationId: "updateStudentById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of student to update"
        required: true
        type: "string"
        format: "hexadecimal"
        example: "5ad289124c9956964c868f2b"
      - name: "body"
        in: "body"
        description: "Student object that needs to be updated. if _id is part of the object it should match exactly the path parameter"
        required: true
        schema:
          $ref: "#/definitions/Student"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Student"
    delete:
      tags:
      - "students"
      summary: "Delete a student by id"
      operationId: "deleteStudentById"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of student to delete"
        required: true
        type: "string"
        format: "hexadecimal"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
...
definitions:
  Student:
    type: "object"
    properties:
      _id:
        type: "string"
        format: "hexadecimal"
        description: "mLab generated mongodb field"
        example: "12312d1231b"
        required: false
      name:
        type: "string"
        format: "First Middle Last"
        description: "Full name of the student"
        example: "Христо Евлогиев Георгиев"
      group:
        type: "string"
        example: "1а"
      avatar:
        type: "string"
        example: "http://www.sdjgjx.com/up/pc/baby cartoon.jpg"
    xml:
      name: "Student"
...